WT-2571 In join tests, use scenarios instead of lots of nested loops. Move miscellaneous tests that don't need scenarios to their own test class.
